Renamed `_config` as `core`
