--- database
@@ -1,25 +1,0 @@
-if [ "$#" -ne 2 ]; then
-    echo "Usage: $0 [flush | rebuild | refresh]"
-    exit 1
-fi
-
-case "$1" in
-    "flush")
-        rm -rv ./*/migrations/0*.py ./*/__pycache__ db.sqlite3
-        ;;
-    "rebuild")
-        sh database flush $2
-        sh database refresh $2
-        python manage.py createsuperuser --settings=_config.settings.$2
-        ;;
-    "refresh")
-        python manage.py makemigrations --settings=_config.settings.$2
-        python manage.py migrate --settings=_config.settings.$2
-        ;;
-    *)
-        echo "Unknown parameter: $1"
-        exit 1
-        ;;
-esac
-
-exit 0
